+++ lib/recording_studio_stripe/cancel_subscription.rb
@@ -2,16 +2,20 @@
 
 module RecordingStudioStripe
   class CancelSubscription
-    def self.call(root_recording:)
-      new(root_recording: root_recording).call
+    def self.call(root_recording:, subscription_type: nil)
+      new(root_recording: root_recording, subscription_type: subscription_type).call
     end
 
-    def initialize(root_recording:)
+    def initialize(root_recording:, subscription_type:)
       @root_recording = root_recording
+      @subscription_type = subscription_type
     end
 
     def call
-      subscription = Subscription.current_for(root_recording_id: @root_recording.id)
+      subscription = Subscription.current_for(
+        root_recording_id: @root_recording.id,
+        subscription_type: @subscription_type.presence
+      )
       raise NoSubscription, "Nothing to cancel" unless subscription
 
       unless RecordingStudioStripe.configuration.local_mode?